+++ lib/mix/lib/mix/tasks/loadpaths.ex
@@ -13,7 +13,7 @@ defmodule Mix.Tasks.Loadpaths do
     {opts, _, _} = OptionParser.parse(args)
 
     unless opts[:no_elixir_version_check] do
-      config = Mix.project
+      config = Mix.Project.config
 
       if req = config[:elixir] do
         case Version.parse_requirement(req) do